+++ ci/raydepsets/BUILD.bazel
@@ -50,13 +50,14 @@ py_test(
     deps = [
         ci_require("pytest"),
         ":raydepsets_lib",
-        ":test_utils",
+        ":utils",
     ],
 )
 
 py_library(
-    name = "test_utils",
-    srcs = ["tests/test_utils.py"],
+    name = "utils",
+    testonly = True,
+    srcs = ["tests/utils.py"],
     deps = [
         ci_require("bazel-runfiles"),
     ],
@@ -74,7 +75,7 @@ py_test(
     ],
     deps = [
         ci_require("pytest"),
-        ":test_utils",
+        ":utils",
         ":workspace",
     ],
 )
