+++ ci/raydepsets/tests/test_workspace.py
@@ -3,7 +3,7 @@ import pytest
 import tempfile
 from pathlib import Path
 from ci.raydepsets.workspace import Workspace
-from ci.raydepsets.tests.test_utils import copy_data_to_tmpdir
+from ci.raydepsets.tests.utils import copy_data_to_tmpdir
 
 
 def test_workspace_init():
